# api_gateway/api/docs/swagger.yaml
definitions:
...
  models.ResponseModel:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
  models.Status:
    properties:
      status:
        type: string
    type: object
...
  order_service.CreateOrderRequest:
    properties:
      address:
        type: string
      branch_id:
        type: integer
      client_id:
        type: integer
      courier_id:
        type: integer
      delivery_price:
        type: number
      discount:
        type: number
      payment_type:
        description: '''cash'',''card'''
        type: string
      price:
        type: number
      type:
        description: delivery, pick-up
        type: string
    type: object
...
  order_service.ListOrderResponse:
    properties:
      Orders:
        items:
          $ref: '#/definitions/order_service.Order'
        type: array
      count:
        type: integer
    type: object
  order_service.Order:
    properties:
      address:
        type: string
      branch_id:
        type: integer
      client_id:
        type: integer
      courier_id:
        type: integer
      created_at:
        type: string
      delivery_price:
        type: number
      discount:
        type: number
      id:
        type: integer
      order_id:
        type: string
      payment_type:
        description: '''cash'',''card'''
        type: string
      price:
        type: number
      status:
        type: string
      type:
        description: delivery, pick-up
        type: string
      updated_at:
        type: string
    type: object
  order_service.Response:
    properties:
      message:
        type: string
    type: object
...
  order_service.UpdateOrderRequest:
    properties:
      address:
        type: string
      branch_id:
        type: integer
      client_id:
        type: integer
      courier_id:
        type: integer
      delivery_price:
        type: number
      discount:
        type: number
      id:
        type: integer
      order_id:
        type: string
      payment_type:
        description: '''cash'',''card'''
        type: string
      price:
        type: number
      status:
        type: string
      type:
        description: delivery, pick-up
        type: string
    type: object
  order_service.UpdateOrderStatusRequest:
    properties:
      id:
        type: integer
      order_id:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/order/create:
    post:
      consumes:
      - application/json
      description: create order
      operationId: create-order
      parameters:
      - description: order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order_service.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/order_service.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: create order
      tags:
      - order
  /v1/order/delete/{order_id}:
    delete:
      consumes:
      - application/json
      description: Delete Order
      operationId: delete-order
      parameters:
      - description: order_id
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.Status'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: delete order
      tags:
      - order
  /v1/order/get/{order_id}:
    get:
      consumes:
      - application/json
      description: get order
      operationId: get-order-byID
      parameters:
      - description: order_id
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/order_service.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get order by ID
      tags:
      - order
  /v1/order/list:
    get:
      consumes:
      - application/json
      description: get order
      operationId: get-order
      parameters:
      - description: search
        in: query
        name: search
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: order_id
        in: query
        name: order_id
        type: integer
      - description: client_id
        in: query
        name: client_id
        type: integer
      - description: branch_id
        in: query
        name: branch_id
        type: integer
      - description: delivery_type
        in: query
        name: delivery_type
        type: string
      - description: courier_id
        in: query
        name: courier_id
        type: integer
      - description: price_from
        in: query
        name: price_from
        type: integer
      - description: price_to
        in: query
        name: price_to
        type: integer
      - description: payment_type
        in: query
        name: payment_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/order_service.ListOrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get list order
      tags:
      - order
  /v1/order/update/{order_id}:
    put:
      consumes:
      - application/json
      description: Update Order by ID
      operationId: update_order
      parameters:
      - description: order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order_service.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.Status'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Update Order
      tags:
      - order
  /v1/order/update/status/{order_id}:
    put:
      consumes:
      - application/json
      description: Update status by ID and status type
      operationId: update_order_status
      parameters:
      - description: order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order_service.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.Status'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Update Order STATUS
      tags:
      - order
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"